+++ R/data.R
@@ -54,7 +54,7 @@
 #' @references {Jarvis, A, HI Reuter, A Nelson, E Guevara, 2008, Hole-filled
 #' SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database
 #' \url{http://srtm.csi.cgiar.org}}
-"stations"
+"GSOD_stations"
 
 #' country_list
 #' @format A data frame with 293 observations of 4 variables:
@@ -78,4 +78,4 @@
 #'
 #' @source \url{ftp://ftp.ncdc.noaa.gov/pub/data/noaa/country-list.txt}
 #'
-"country_list"
+"GSOD_country_list"